fix: corrige aprovacao de produtos e configura Java 17 permanente
- Corrige tipo do ID de Long para Integer no ProdutoRepository e ProdutoController
  (bug que impedia aprovacao de produtos no painel admin)
- Corrige campo condicao -> estado no frontend (Home.js)
- Adiciona .devcontainer/devcontainer.json com Java 17 para garantir
  ambiente correto automaticamente ao abrir o projeto no Codespaces

+++ backend/src/main/java/com/doacaobebe/controller/ProdutoController.java
@@ -77,7 +77,7 @@ public class ProdutoController {
     }
 
     @PutMapping("/{id}/aprovar")
-    public ResponseEntity<String> aprovarProduto(@PathVariable Long id) {
+    public ResponseEntity<String> aprovarProduto(@PathVariable Integer id) {
         Produto produto = produtoRepository.findById(id).orElse(null);
 
         if (produto == null) {
@@ -91,7 +91,7 @@ public class ProdutoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> removerProduto(@PathVariable Long id) {
+    public ResponseEntity<String> removerProduto(@PathVariable Integer id) {
         produtoRepository.deleteById(id);
         return ResponseEntity.ok("Produto removido com sucesso!");
     }
